[Feat] domain, controller 나누어 구현
